Remove default split mode from db schema
There was a mismatch: In the db the default was 0 (mine) but in the code
it's undecided (2). I don't see much value in setting it in the db and
this can cause issues if I forget to update it in both places again, so
I'd rather just have it in the code.