Destroy demo user when demo expires
Instead of deleting the demo user when the session is deleted (which
only happens if the user manually logs out) we now schedule a job to
delete the user after 1 hour, as soon as the demo user is created.

I also moved everthing related to demo users in a new User::Demoable
concern.