Force SSL in production
This was set to true by default, then I changed it to false when setting
up kamal. I must have had a reason at the time, but it seems wrong to
have it as false.

Plus, it might be causing intermittent 422 errors on login due to stale
CSRF tokens (long story).